Make if statements more condensed

+++ upload.sh
@@ -1,15 +1,6 @@
 #!/bin/bash
 
-if [ "$#" -ne 1 ] || ! [ -d "$1" ]
-then
-    echo "Usage: $0 UPLOAD_DIR"
-    exit 1
-fi
-
-if [ ! -f $OUT/lineage-* ]
-then
-    echo "Error: No files to upload"
-    exit 1
-fi
+[ "$#" == 1 ] || [ -d "$1" ] || { echo "Usage: $0 UPLOAD_DIR"; exit 1; }
+[ -f $OUT/lineage-* ] || { echo "Error: No files to upload"; exit 1; }
 
 mv $OUT/lineage-* $1
